Able to have the maintasklist component be be integrated with the profile card, we are able to link each task to their job card, next is to style the profile card and finish up the task component to have delete and edit
